// kaghati-shopify-v3/.gadget/client/src/types.ts
type JSONValue =
    | string
    | number
    | boolean
    | JSONObject
    | JSONArray;

interface JSONObject {
    [x: string]: JSONValue;
}
// ...
export type SortOrder = "Ascending"|"Descending";
// ...
export type ShopifyBulkOperationErrorCodeEnum = "access_denied" | "internal_server_error" | "timeout";

/** Represents the possible values of the Status enum. */
export type ShopifyBulkOperationStatusEnum = "canceled" | "canceling" | "completed" | "created" | "expired" | "failed" | "running";

/** Represents the possible values of the Type enum. */
export type ShopifyBulkOperationTypeEnum = "query" | "mutation";
// ...
export type Upload = any;
// ...
export type IDFilter = {

  equals?: (Scalars['GadgetID'] | null) | null;

  notEquals?: (Scalars['GadgetID'] | null) | null;

  isSet?: (Scalars['Boolean'] | null) | null;

  in?: ((Scalars['GadgetID'] | null) | null)[];

  notIn?: ((Scalars['GadgetID'] | null) | null)[];

  lessThan?: (Scalars['GadgetID'] | null) | null;

  lessThanOrEqual?: (Scalars['GadgetID'] | null) | null;

  greaterThan?: (Scalars['GadgetID'] | null) | null;

  greaterThanOrEqual?: (Scalars['GadgetID'] | null) | null;
};



export type DateTimeFilter = {

  equals?: Date | Scalars['ISO8601DateString'] | null;

  notEquals?: Date | Scalars['ISO8601DateString'] | null;

  isSet?: (Scalars['Boolean'] | null) | null;

  in?: (Date | Scalars['ISO8601DateString'] | null)[];

  notIn?: (Date | Scalars['ISO8601DateString'] | null)[];

  lessThan?: Date | Scalars['ISO8601DateString'] | null;

  lessThanOrEqual?: Date | Scalars['ISO8601DateString'] | null;

  greaterThan?: Date | Scalars['ISO8601DateString'] | null;

  greaterThanOrEqual?: Date | Scalars['ISO8601DateString'] | null;

  before?: Date | Scalars['ISO8601DateString'] | null;

  after?: Date | Scalars['ISO8601DateString'] | null;
};
// ...
export type StringFilter = {

  equals?: (Scalars['String'] | null) | null;

  notEquals?: (Scalars['String'] | null) | null;

  isSet?: (Scalars['Boolean'] | null) | null;

  in?: ((Scalars['String'] | null) | null)[];

  notIn?: ((Scalars['String'] | null) | null)[];

  lessThan?: (Scalars['String'] | null) | null;

  lessThanOrEqual?: (Scalars['String'] | null) | null;

  greaterThan?: (Scalars['String'] | null) | null;

  greaterThanOrEqual?: (Scalars['String'] | null) | null;

  startsWith?: (Scalars['String'] | null) | null;
};
// ...
export type SingleEnumFilter = {

  isSet?: (Scalars['Boolean'] | null) | null;

  equals?: (Scalars['String'] | null) | null;

  notEquals?: (Scalars['String'] | null) | null;

  in?: ((Scalars['String'] | null) | null)[];
};
// ...
export type StateFilter = {

  isSet?: (Scalars['Boolean'] | null) | null;

  inState?: (Scalars['String'] | null) | null;
};
// ...
export type ShopifyBulkOperationSort = {

  /** Sort the results by the id field. Defaults to ascending (smallest value first). */
  id?: SortOrder | null;

  /** Sort the results by the createdAt field. Defaults to ascending (smallest value first). */
  createdAt?: SortOrder | null;

  /** Sort the results by the updatedAt field. Defaults to ascending (smallest value first). */
  updatedAt?: SortOrder | null;

  /** Sort the results by the state field. Defaults to ascending (smallest value first). */
  state?: SortOrder | null;

  /** Sort the results by the completedAt field. Defaults to ascending (smallest value first). */
  completedAt?: SortOrder | null;

  /** Sort the results by the errorCode field. Defaults to ascending (smallest value first). */
  errorCode?: SortOrder | null;

  /** Sort the results by the fileSize field. Defaults to ascending (smallest value first). */
  fileSize?: SortOrder | null;

  /** Sort the results by the objectCount field. Defaults to ascending (smallest value first). */
  objectCount?: SortOrder | null;

  /** Sort the results by the partialDataUrl field. Defaults to ascending (smallest value first). */
  partialDataUrl?: SortOrder | null;

  /** Sort the results by the query field. Defaults to ascending (smallest value first). */
  query?: SortOrder | null;

  /** Sort the results by the rootObjectCount field. Defaults to ascending (smallest value first). */
  rootObjectCount?: SortOrder | null;

  /** Sort the results by the status field. Defaults to ascending (smallest value first). */
  status?: SortOrder | null;

  /** Sort the results by the type field. Defaults to ascending (smallest value first). */
  type?: SortOrder | null;

  /** Sort the results by the url field. Defaults to ascending (smallest value first). */
  url?: SortOrder | null;
};



export type ShopifyBulkOperationFilter = {

  AND?: (ShopifyBulkOperationFilter | null)[];

  OR?: (ShopifyBulkOperationFilter | null)[];

  NOT?: (ShopifyBulkOperationFilter | null)[];

  id?: IDFilter | null;

  createdAt?: DateTimeFilter | null;

  updatedAt?: DateTimeFilter | null;

  state?: StateFilter | null;

  completedAt?: DateTimeFilter | null;

  errorCode?: SingleEnumFilter | null;

  fileSize?: StringFilter | null;

  objectCount?: StringFilter | null;

  partialDataUrl?: StringFilter | null;

  query?: StringFilter | null;

  rootObjectCount?: StringFilter | null;

  status?: SingleEnumFilter | null;

  type?: SingleEnumFilter | null;

  url?: StringFilter | null;

  shop?: IDFilter | null;

  shopId?: IDFilter | null;
};
// ...
export interface Scalars {
  /** Represents an amount of some currency. Specified as a string so user's aren't tempted to do math on the value. */
  CurrencyAmount: string;
  /** Represents a UTC date formatted an ISO-8601 formatted 'full-date' string. */
  ISO8601DateString: string;
  /** A file object produced by a browser for uploading to cloud storage */
  Upload: File;
  /** The `String` scalar type represents textual data, represented as UTF-8 character sequences. The String type is most often used by GraphQL to represent free-form human-readable text. */
  String: string;
  /** The `JSONObject` scalar type represents JSON objects as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSONObject: JSONObject;
  /** The `Boolean` scalar type represents `true` or `false`. */
  Boolean: boolean;
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: JSONValue;
  /** The ID of a record in Gadget */
  GadgetID: string;
  /** A date-time string at UTC, such as 2007-12-03T10:15:30Z, compliant with the `date-time` format outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representation of dates and times using the Gregorian calendar. */
  DateTime: Date;
  /** A field whose value conforms to the standard URL format as specified in RFC3986: https://www.ietf.org/rfc/rfc3986.txt. */
  URL: string;
  /** A field whose value conforms to the standard internet email address format as specified in RFC822: https://www.w3.org/Protocols/rfc822/. */
  EmailAddress: string;
  /** Represents the state of one record in a Gadget database. Represented as either a string or set of strings nested in objects. */
  RecordState: (string | { [key: string]: Scalars['RecordState'] });
  /** The `Float` scalar type represents signed double-precision fractional values as specified by [IEEE 754](https://en.wikipedia.org/wiki/IEEE_floating_point). */
  Float: number;
  /** The `Int` scalar type represents non-fractional signed whole numeric values. Int can represent values between -(2^31) and 2^31 - 1. */
  Int: number;
  /** Instructions for a client to turn raw transport types (like strings) into useful client side types (like Dates). Unstable and not intended for developer use. */
  HydrationPlan: unknown;
};
// ...
export type Query = {

  __typename: 'Query';

  user: (User | null);

  users: UserConnection;

  session: (Session | null);

  sessions: SessionConnection;

  shopifyAsset: (ShopifyAsset | null);

  shopifyAssets: ShopifyAssetConnection;

  shopifyBulkOperation: (ShopifyBulkOperation | null);
// ...
};
// ...
export type ShopifyBulkOperation = {

  __typename: 'ShopifyBulkOperation';

  /** The globally unique, unchanging identifier for this record. Assigned and managed by Shopify. */
  id: Scalars['GadgetID'];

  /** The time at which this record was first created. Set once upon record creation and never changed. Managed by Gadget. */
  createdAt: Scalars['DateTime'];

  /** The time at which this record was last changed. Set each time the record is successfully acted upon by an action. Managed by Gadget. */
  updatedAt: Scalars['DateTime'];

  /** The current state this record is in. Changed by invoking actions. Managed by Gadget. */
  state: Scalars['RecordState'];

  completedAt: (Scalars['DateTime'] | null);

  errorCode: ShopifyBulkOperationErrorCodeEnum;

  fileSize: (Scalars['String'] | null);

  objectCount: (Scalars['String'] | null);

  partialDataUrl: (Scalars['String'] | null);

  query: (Scalars['String'] | null);

  rootObjectCount: (Scalars['String'] | null);

  status: ShopifyBulkOperationStatusEnum;

  type: ShopifyBulkOperationTypeEnum;

  url: (Scalars['String'] | null);

  shop: ShopifyShop;

  shopId: Scalars['GadgetID'];

  /** Get all the fields for this record. Useful for not having to list out all the fields you want to retrieve, but slower. */
  _all: Scalars['JSONObject'];
};



export type AvailableShopifyBulkOperationSelection = {

  __typename?: boolean | null | undefined;

  /** The globally unique, unchanging identifier for this record. Assigned and managed by Shopify. */
  id?: boolean | null | undefined;

  /** The time at which this record was first created. Set once upon record creation and never changed. Managed by Gadget. */
  createdAt?: boolean | null | undefined;

  /** The time at which this record was last changed. Set each time the record is successfully acted upon by an action. Managed by Gadget. */
  updatedAt?: boolean | null | undefined;

  /** The current state this record is in. Changed by invoking actions. Managed by Gadget. */
  state?: boolean | null | undefined;

  completedAt?: boolean | null | undefined;

  errorCode?: boolean | null | undefined;

  fileSize?: boolean | null | undefined;

  objectCount?: boolean | null | undefined;

  partialDataUrl?: boolean | null | undefined;

  query?: boolean | null | undefined;

  rootObjectCount?: boolean | null | undefined;

  status?: boolean | null | undefined;

  type?: boolean | null | undefined;

  url?: boolean | null | undefined;

  shop?: AvailableShopifyShopSelection;

  shopId?: boolean | null | undefined;

  /** Get all the fields for this record. Useful for not having to list out all the fields you want to retrieve, but slower. */
  _all?: boolean | null | undefined;
};
// ...
export type DeepFilterNever<T> = T extends Record<string, unknown> ? FilterNever<{
  [Key in keyof T]: T[Key] extends Record<string, unknown> ? DeepFilterNever<T[Key]> : T[Key];
}> : T;

export type Select<Schema, Selection extends FieldSelection | null | undefined> = Selection extends null | undefined
  ? never
  : Schema extends (infer T)[]
  ? Select<T, Selection>[]
  : Schema extends null
  ? Select<Exclude<Schema, null>, Selection> | null
  : {
      [Key in keyof Selection & keyof Schema]: Selection[Key] extends true
        ? Schema[Key]
        : Selection[Key] extends FieldSelection
        ? Select<Schema[Key], Selection[Key]>
        : never;
    };